исправлены теги и возможность удалять чужие полки, книги, теги

+++ local/modules/up.litlab/install/components/up/bookshelf.edit/ajax.php
@@ -1,5 +1,6 @@
 <?php
 
+use Bitrix\Main\Context;
 use Bitrix\Main\DI\ServiceLocator;
 use Up\Litlab\API\Bookshelf;
 
@@ -31,27 +32,39 @@ class BookshelfEditAjaxController extends \Bitrix\Main\Engine\Controller
 
 	public function deleteBookAction($bookId, $bookshelfId, $action)
 	{
-		if (!$_SESSION['NAME'] || !$bookshelfId || !$bookId || !$action)
+		$bookshelfApi = ServiceLocator::getInstance()->get('Bookshelf');
+		if (!$_SESSION['USER_ID'] || !$bookshelfId || !$bookId || !$action)
 		{
 			return ['result' => false];
 		}
-		if($action === 'delete')
+
+		if($_SESSION['USER_ID'] !== $bookshelfApi->getBookshelfCreator($bookshelfId)){
+			return ['result' => false];
+		}
+
+		if(($action === 'delete') && $_SESSION['USER_ID'] === $bookshelfApi->getBookshelfCreator($bookshelfId))
 		{
-			$bookshelfApi = ServiceLocator::getInstance()->get('Bookshelf');
 			$bookshelfApi->deleteBookOfBookshelf([$bookId], $bookshelfId);
+
 			return ['result' => true];
 		}
 	}
 
 	public function deleteBookshelfAction($bookshelfId, $action)
 	{
-		if (!$_SESSION['NAME'] || !$bookshelfId || !$action)
+		if (!$_SESSION['USER_ID'] || !$bookshelfId || !$action)
 		{
 			return ['result' => false];
 		}
-		if($action === 'delete')
+
+		$bookshelfApi = ServiceLocator::getInstance()->get('Bookshelf');
+
+		if($_SESSION['USER_ID'] !== $bookshelfApi->getBookshelfCreator($bookshelfId)){
+			return ['result' => false];
+		}
+
+		if($action === 'delete' && $_SESSION['USER_ID'] === $bookshelfApi->getBookshelfCreator($bookshelfId))
 		{
-			$bookshelfApi = ServiceLocator::getInstance()->get('Bookshelf');
 			$bookshelfApi->updateStatus($bookshelfId, 'deleted');
 			return ['result' => true];
 		}
@@ -60,49 +73,23 @@ class BookshelfEditAjaxController extends \Bitrix\Main\Engine\Controller
 
 	public function deleteTagAction($bookshelfId, $tagId, $action)
 	{
-		if (!$_SESSION['NAME'] || !$bookshelfId || !$tagId || !$action)
+		if (!$_SESSION['USER_ID'] || !$bookshelfId || !$tagId || !$action)
 		{
 			return ['result' => false];
 		}
+		$bookshelfApi = ServiceLocator::getInstance()->get('Bookshelf');
+
+		if($_SESSION['USER_ID'] !== $bookshelfApi->getBookshelfCreator($bookshelfId)){
+			return ['result' => false];
+		}
+
 		if($action === 'delete')
 		{
-			$bookshelfApi = ServiceLocator::getInstance()->get('Bookshelf');
 			$bookshelfApi->deleteTagOfBookshelf((int)$tagId, (int)$bookshelfId);
-			return ['result' => true, 'action'=>'delete'];
+			return ['result' => true];
 
 		}
 		return ['result'=>false];
 	}
 
-	public function addTagAction($bookshelfId, $tag, $action)
-	{
-		$bookshelfApi = ServiceLocator::getInstance()->get('Bookshelf');
-		var_dump($tag);
-		if($action==='addTag')
-		{
-			if ($bookshelfApi->getTagByName($tag) === false) //если тега не существует
-			{
-				$bookshelfApi->addTag($tag);
-				$bookshelfApi->addTagsOfBookshelf((int)$bookshelfApi->getTagByName($tag)['ID'], (int)$bookshelfId);
-			}
-			else
-			{
-				if (!in_array($tag, $bookshelfApi->getTags($bookshelfId)))
-				{ // если существует и такой связи нет
-					$bookshelfApi->addTagsOfBookshelf(
-						(int)$bookshelfApi->getTagByName($tag)['ID'],
-						(int)$bookshelfId
-					);
-				}
-			}
-
-			return ['result' => true, 'action' => 'addTag'];
-		}
-		return ['result'=>false];
-	}
-
-	public function editTagValueAction()
-	{
-
-	}
 }
